# runs/WI-001/solution-architect/supporting-files/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LittleDuck MVP API
  version: 1.0.0
  x-contract-revision: 7
  description: |
    WI-001 revision 7 third-review baseline API contract.
    All response and event timestamps are serialized as UTC ISO 8601 with a
    zero offset. User-facing grouping and admin date
    filters use Asia/Shanghai semantics as documented on each operation.
    Semantic changes require a Coordinator-approved contract revision Work Item.
servers:
  - url: /
    description: Same-origin production and local reverse proxy
...
paths:
...
  /api/v1/user/auth/logout:
    post:
      tags: [User Auth]
      operationId: logoutUser
      summary: Revoke the current user session
      description: |
        Active generations started by this exact user session receive a stop
        request. Other sessions for the same user are not revoked.
      security:
        - UserCookie: []
      responses:
        "204":
          description: Session revoked and cookie cleared
          headers:
            Set-Cookie:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/user/generations:
    post:
      tags: [User Generations]
      operationId: createUserGeneration
      summary: Persist a user message and stream a new assistant generation
      description: |
        conversationId is omitted for a new unsaved conversation. The first
        successful persistence transaction creates the conversation, temporary
        title, user message, assistant placeholder, generation and LLM call.
        clientMessageId is generated once by the browser. Repeating it for the
        same user returns 409 DUPLICATE_MESSAGE with the existing generationId,
        so the client can read authoritative state without creating another message.
      security:
        - UserCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGenerationRequest"
            examples:
              newConversation:
                value:
                  clientMessageId: d5026980-61a4-4478-a045-bd2efbf17abc
                  content: 帮我写一个 Python 脚本
              existingConversation:
                value:
                  conversationId: 6bf22123-6db0-4232-929d-3c77272a6ad2
                  clientMessageId: 4a055129-4e14-4624-931f-8f4661bd1232
                  content: 这个脚本如何增加参数校验？
      responses:
        "200":
          description: |
            SSE stream. The first event is generation.started, followed by zero
            or more generation.delta events and exactly one terminal event.
          content:
            text/event-stream:
              schema:
                type: string
                format: binary
              example: |
                event: generation.started
                data: {"generationId":"2c255ae2-d891-4d6c-80c5-3d5ee1f534ca","sequence":1,"kind":"chat","conversationId":"6bf22123-6db0-4232-929d-3c77272a6ad2","userMessageId":"d5026980-61a4-4478-a045-bd2efbf17abc","assistantMessageId":"4cbe9c70-74ff-4d5c-a70f-131722b21589","temporaryTitle":"帮我写一个 Python 脚本","occurredAt":"2026-07-16T12:00:00Z"}

                event: generation.delta
                data: {"generationId":"2c255ae2-d891-4d6c-80c5-3d5ee1f534ca","sequence":2,"assistantMessageId":"4cbe9c70-74ff-4d5c-a70f-131722b21589","delta":"当然可以。","accumulatedLength":6,"occurredAt":"2026-07-16T12:00:01Z"}

          x-sse-events:
            - $ref: "#/components/schemas/GenerationStartedEvent"
            - $ref: "#/components/schemas/GenerationDeltaEvent"
            - $ref: "#/components/schemas/GenerationCompletedEvent"
            - $ref: "#/components/schemas/GenerationFailedEvent"
            - $ref: "#/components/schemas/GenerationStoppedEvent"
            - $ref: "#/components/schemas/HeartbeatEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Active generation or duplicate clientMessageId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          description: LLM configuration is missing or service unavailable before persistence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/user/generations/{generationId}:
    get:
      tags: [User Generations]
      operationId: getUserGeneration
      summary: Read the authoritative generation state
      security:
        - UserCookie: []
      parameters:
        - $ref: "#/components/parameters/GenerationId"
      responses:
        "200":
          description: Generation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/generations/{generationId}/stop:
    post:
      tags: [User Generations]
      operationId: stopUserGeneration
      summary: Idempotently request generation stop
      security:
        - UserCookie: []
      parameters:
        - $ref: "#/components/parameters/GenerationId"
      responses:
        "200":
          description: Generation was already terminal; current state returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "202":
          description: Stop accepted; observe SSE or generation state for terminal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/assistant-messages/{assistantMessageId}/retries:
    post:
      tags: [User Generations]
      operationId: retryAssistantMessage
      summary: Retry the latest retryable failed or stopped assistant message
      description: |
        Reuses the original user message, creates a new assistant message,
        generation and LLM call, and preserves the original assistant message.
      security:
        - UserCookie: []
      parameters:
        - $ref: "#/components/parameters/AssistantMessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryGenerationRequest"
      responses:
        "200":
          description: SSE stream with kind=retry
          content:
            text/event-stream:
              schema:
                type: string
                format: binary
          x-sse-events:
            - $ref: "#/components/schemas/GenerationStartedEvent"
            - $ref: "#/components/schemas/GenerationDeltaEvent"
            - $ref: "#/components/schemas/GenerationCompletedEvent"
            - $ref: "#/components/schemas/GenerationFailedEvent"
            - $ref: "#/components/schemas/GenerationStoppedEvent"
            - $ref: "#/components/schemas/HeartbeatEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Message is no longer retryable or another generation is active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: RETRY_NOT_ALLOWED
                  message: 该回复当前不可重试
                  requestId: req_01
...
components:
  securitySchemes:
    UserCookie:
      type: apiKey
      in: cookie
      name: ld_user_session
      description: HttpOnly user session cookie scoped to /api/v1/user
    AdminCookie:
      type: apiKey
      in: cookie
      name: ld_admin_session
      description: HttpOnly admin session cookie scoped to /api/v1/admin
  parameters:
...
    GenerationId:
      name: generationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssistantMessageId:
      name: assistantMessageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid input, verification code, page or date range
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: 请求参数不正确
                  requestId: req_01
                  details:
                    - field: content
                      issue: 长度必须为 1 到 4000 个字符
            invalidCode:
              value:
                error:
                  code: INVALID_VERIFICATION_CODE
                  message: 验证码错误，请重新输入
                  requestId: req_01
    Unauthorized:
      description: Session is absent, expired or revoked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: UNAUTHENTICATED
              message: 登录状态已失效，请重新登录
              requestId: req_01
    Forbidden:
      description: Same-origin or authorization validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: FORBIDDEN
              message: 请求未获授权
              requestId: req_01
    NotFound:
      description: Resource does not exist or is not owned by the current user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在
              requestId: req_01
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the client may retry
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: RATE_LIMITED
              message: 请求过于频繁，请稍后再试
              requestId: req_01
  schemas:
...
    Message:
      type: object
      additionalProperties: false
      required: [id, conversationId, sequence, role, status, content, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 1
          description: Stable, strictly increasing order inside one conversation
        role:
          type: string
          enum: [user, assistant]
        status:
          type: string
          enum: [persisted, generating, completed, failed, stopped]
        content:
          type: string
        replyToMessageId:
          type: string
          format: uuid
        retryOfMessageId:
          type: string
          format: uuid
        generationId:
          type: string
          format: uuid
        canRetry:
          type: boolean
          description: Present for assistant messages
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateGenerationRequest:
      type: object
      additionalProperties: false
      required: [clientMessageId, content]
      properties:
        conversationId:
          type: string
          format: uuid
          description: Omit for an unsaved new conversation
        clientMessageId:
          type: string
          format: uuid
          description: Stable browser-generated ID for duplicate prevention
        content:
          type: string
          minLength: 1
          maxLength: 4000
          description: Server trims surrounding whitespace before applying the length limits
    RetryGenerationRequest:
      type: object
      additionalProperties: false
      required: [clientRetryId]
      properties:
        clientRetryId:
          type: string
          format: uuid
          description: Stable browser-generated ID used to prevent duplicate retries
    Generation:
      type: object
      additionalProperties: false
      required:
        - id
        - conversationId
        - userMessageId
        - assistantMessageId
        - kind
        - status
        - stopRequested
        - errorCode
        - startedAt
        - finishedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        userMessageId:
          type: string
          format: uuid
        assistantMessageId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [chat, retry]
        status:
          type: string
          enum: [streaming, completed, failed, stopped]
        stopRequested:
          type: boolean
        errorCode:
          type: string
          nullable: true
          enum:
            - LLM_NOT_CONFIGURED
            - LLM_UNAVAILABLE
            - LLM_RATE_LIMITED
            - LLM_TIMEOUT
            - RESPONSE_PARSE_FAILED
            - GENERATION_INTERRUPTED
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GenerationResponse:
      type: object
      additionalProperties: false
      required: [generation, assistantMessage]
      properties:
        generation:
          $ref: "#/components/schemas/Generation"
        assistantMessage:
          $ref: "#/components/schemas/Message"
    GenerationStartedEvent:
      type: object
      additionalProperties: false
      required:
        - generationId
        - sequence
        - kind
        - conversationId
        - userMessageId
        - assistantMessageId
        - temporaryTitle
        - occurredAt
      properties:
        generationId:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 1
        kind:
          type: string
          enum: [chat, retry]
        conversationId:
          type: string
          format: uuid
        userMessageId:
          type: string
          format: uuid
        assistantMessageId:
          type: string
          format: uuid
        temporaryTitle:
          type: string
          maxLength: 20
        occurredAt:
          type: string
          format: date-time
    GenerationDeltaEvent:
      type: object
      additionalProperties: false
      required:
        - generationId
        - sequence
        - assistantMessageId
        - delta
        - accumulatedLength
        - occurredAt
      properties:
        generationId:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 2
        assistantMessageId:
          type: string
          format: uuid
        delta:
          type: string
          minLength: 1
        accumulatedLength:
          type: integer
          minimum: 1
        occurredAt:
          type: string
          format: date-time
    GenerationCompletedEvent:
      type: object
      additionalProperties: false
      required:
        - generationId
        - sequence
        - generation
        - assistantMessage
        - titleWillBeAttempted
        - occurredAt
      properties:
        generationId:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 2
        generation:
          $ref: "#/components/schemas/Generation"
        assistantMessage:
          $ref: "#/components/schemas/Message"
        titleWillBeAttempted:
          type: boolean
          description: True when the service will attempt the non-blocking first-title call
        occurredAt:
          type: string
          format: date-time
    GenerationFailedEvent:
      type: object
      additionalProperties: false
      required:
        [generationId, sequence, generation, assistantMessage, error, occurredAt]
      properties:
        generationId:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 2
        generation:
          $ref: "#/components/schemas/Generation"
        assistantMessage:
          $ref: "#/components/schemas/Message"
        error:
          $ref: "#/components/schemas/PublicGenerationError"
        occurredAt:
          type: string
          format: date-time
    GenerationStoppedEvent:
      type: object
      additionalProperties: false
      required:
        [generationId, sequence, generation, assistantMessage, stoppedBy, occurredAt]
      properties:
        generationId:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 2
        generation:
          $ref: "#/components/schemas/Generation"
        assistantMessage:
          $ref: "#/components/schemas/Message"
        stoppedBy:
          type: string
          enum: [user, logout]
        occurredAt:
          type: string
          format: date-time
    HeartbeatEvent:
      type: object
      additionalProperties: false
      required: [generationId, occurredAt]
      properties:
        generationId:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
    PublicGenerationError:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code:
          type: string
          enum:
            - LLM_NOT_CONFIGURED
            - LLM_UNAVAILABLE
            - LLM_RATE_LIMITED
            - LLM_TIMEOUT
            - RESPONSE_PARSE_FAILED
            - GENERATION_INTERRUPTED
        message:
          type: string
        retryable:
          type: boolean
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [issue]
      properties:
        field:
          type: string
        issue:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - INVALID_VERIFICATION_CODE
            - USER_ALREADY_EXISTS
            - USER_NOT_REGISTERED
            - INVALID_CREDENTIALS
            - UNAUTHENTICATED
            - FORBIDDEN
            - RESOURCE_NOT_FOUND
            - RATE_LIMITED
            - PAGE_INVALID
            - DATE_RANGE_INVALID
            - UNSUPPORTED_MEDIA_TYPE
            - GENERATION_IN_PROGRESS
            - RETRY_NOT_ALLOWED
            - DUPLICATE_MESSAGE
            - LLM_NOT_CONFIGURED
            - LLM_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
        requestId:
          type: string
        generationId:
          type: string
          format: uuid
          description: Present for DUPLICATE_MESSAGE so the client can read existing state
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"